==> 03_Exploratory_analysis.Rmd <==
---
title: " 02 Exploratory analysis before imputation"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse) # Основной пакет
library(readxl) # для чтения файлов экселя xls xlsx
library(skimr)  # для описательной статистики
library(writexl)  # для записи файлов экселя xls xlsx
library(corrr) # для составления корреляционной матрицы
library(pheatmap) # для постройки хитмэпа
library(flextable) # для красивых таблиц
library(ggExtra) # для соединения графиков
library(ggpubr) # для соединения графиков
library(GGally)
library(corrplot)
library(mice)
library(lme4)
library(broom.mixed)
library(BlandAltmanLeh) # для Bland-Altman plots

library(gtsummary) # для описательной статистики


theme_custom <- theme(
    panel.background = element_rect(fill = "white"),
    plot.title = element_text(size = 30, hjust = 0.5),
    plot.subtitle = element_text(size = 25, hjust = 0.5),
    strip.text = element_text(size = 20),
    axis.text = element_text(size = 20),
    axis.title = element_text(size = 25),
    legend.title = element_text(size = 25),
    legend.text = element_text(size = 20)
  )
```

## Download dataset
```{r Download dataset}
vis_all <- readRDS("data/raw/vis_all.rds")
```


## выбор переменных
```{r chose vars}
# выбираем переменные 




data <-
  vis_all %>% select(starts_with(
    c("id",
      "smoking" ,
      "alcohol" ,
      "height" ,
      "weight" ,
      "bmi",
      "sbp_" ,
      "dbp_" ,
      "sad",
      "pulse" ,
      "cholesterol" ,
      "hdl" ,
      "athero_index"
      # "mmse" ,
      # "moca" ,
      # "fab" ,
      # "stsutiational" ,
      # "stpersonal" ,
      # "symboldigit" ,
      # "bdi"
    )
  )) 

data_long <- data %>% 
  pivot_longer(
    cols = -c(id),
    names_pattern = "(.+)_(.+)",
    names_to = c(".value", "id_point")
  )


```



## Описательная статистика.
```{r}



fltab_factor <-
  data %>% 
  select(is.factor) %>% 
  skim() %>% 
  select(skim_variable, n_missing, factor.n_unique, factor.top_counts) %>% 
  flextable() %>% 
  theme_box() %>%
  align(align = "center", part = "all")

fltab_factor %>%  save_as_html( values = NULL, path ="data/raw/fltab_factor.html", lang = "en", title = "&#32;")


fltab_numeric <-
  data %>%
  select( is.numeric & -id) %>%
  skim_without_charts() %>%
  select(
    skim_variable,
    n_missing,
    numeric.mean,
    numeric.sd,
    numeric.p50,
    numeric.p25,
    numeric.p75,
    numeric.p0,
    numeric.p100
  ) %>% rename(
    Mean = numeric.mean,
    Sd = numeric.sd,
    Median = numeric.p50,
    Q1 = numeric.p25,
    Q3 = numeric.p75,
    Min = numeric.p0,
    Max = numeric.p100
  ) %>%
  flextable() %>%
  theme_box() %>%
  align(align = "center", part = "all")

fltab_numeric %>%  save_as_html( values = NULL, path ="data/raw/fltab_numeric.html", lang = "en", title = "&#32;")


fltab_factor
fltab_numeric
```
##  боксплоты
```{r boxplot_before_imp, fig.height=8, fig.width=12}
# Бокс-плоты



vars_for_box <- c( 
            "height" ,
      "weight" ,
      "bmi",
      "sbp" ,
      "dbp" ,
      "sad",
      "pulse" ,
      "cholesterol" ,
      "hdl" ,
      "athero_index")


# boxplot_before_imp_height <-
#   data_long %>%
#   filter(!is.na(height)) %>%
#   ggplot(aes(x = id_point, y = height)) +
#   geom_line(aes(group = id),
#             col = "grey50",
#             alpha = 0.4) +
#   theme_custom  +
#   geom_boxplot(
#     fill = "#DEB0D0",
#     alpha = 0.8,
#     outlier.size = 0.5  ) +
#   labs(title = "height dynamics by visit before data imputation")
# 



fun_boxplot_before_imp <- function(df, mapping, num_var) {
  
  box_plot <- ggplot(df %>%
  filter(!is.na(num_var)),  mapping ) +
    geom_line(aes(y = .data[[num_var]], group = id),
            col = "grey50",
            alpha = 0.4) +
    geom_boxplot(aes(y = .data[[num_var]]),fill = "#DEB0D0",
    alpha = 0.8,
    outlier.size = 0.5  )+
   theme_custom  +
  labs(title = paste( num_var, "dynamics by visit before data imputation"))

return(box_plot) 
}

vars_for_box %>% 
    map(function(num_var) fun_boxplot_before_imp(data_long,  aes(x = id_point),  num_var)) -> List_boxplot_before_imp

List_boxplot_before_imp    




```




## Критерий Мкнеймара

```{r}
# mcnemar.test()
# Сравним различия между тестами для оценки состояния когнитивной сферы  / тесты для оценки эмоционально-аффективной сферы(дихотомические переменные) и жалобами на снижение памяти.
# Для этого заведем новые перемееные переводящие резкльтаты психологических тестов в дихотомические переменные. 1- есть нарушение, 0 - нет нарушений.


dataset_for_mcnemar <- vis_all %>%
  select(
    "id",
    "memoryloss_v8",
    "moca_v8",
    "mmse_v8",
    "fab_v8",
    "symboldigit_v8",
    "bdi_v8",
    "stsutiational_v8" ,
    "stpersonal_v8",
    "moca_bin",
    "mmse_bin",
    "fab_bin" ,
    "symboldigit_bin",
    "bdi_bin",
    "stsutiational_bin",
    "stpersonal_bin"
  )

# dataset_for_mcnemar <- dataset_for_mcnemar %>% 
#   mutate(
#     memoryloss_v8 = ifelse(memoryloss_v8==2, 0, 1 ),
#     moca_bin = ifelse(moca_v8<26, 1,0 ),
#     mmse_bin = ifelse(mmse_v8<29, 1,0 ),
#     fab_bin = ifelse(fab_v8<16, 1,0 ),
#     symboldigit_bin = ifelse(symboldigit_v8<45, 1,0 ),
#     bdi_bin = ifelse(bdi_v8>13, 1,0 ),
#     stsutiational_bin = ifelse(stsutiational_v8>30, 1,0 ),
#     stpersonal_bin = ifelse(stpersonal_v8>30, 1,0 )
#   )



var_for_mcnemar <- c("moca_bin",
"mmse_bin",
"fab_bin" ,
"symboldigit_bin",
"bdi_bin",
"stsutiational_bin",
"stpersonal_bin")

run_mcnemar_test <- function(var) {
  tbl <- table(dataset_for_mcnemar$memoryloss_v8, dataset_for_mcnemar$moca_bin)
  test_result <- mcnemar.test(tbl)
  return(data.frame(variable = var, p.value = round(test_result$p.value, 4) , statistic = round(test_result$statistic, 4)))
}

results_mcmemar <-  map_dfr(var_for_mcnemar, run_mcnemar_test) %>% flextable() 

#  mcnemar.test(table(dataset_for_mcnemar$moca_bin, dataset_for_mcnemar$fab_bin))
```


```{r}
#  GTP  подсказал код 
# Генерируем список бинарных переменных
var_for_mcnemar <- c(
  "memoryloss_v8",
  "moca_bin",
  "mmse_bin",
  "fab_bin",
  "symboldigit_bin",
  "bdi_bin",
  "stsutiational_bin",
  "stpersonal_bin"
)

# Функция для выполнения теста Мак-Нимара для пары переменных
run_mcnemar_test_pair <- function(var1, var2) {
  # Создание таблицы сопряженности
  tbl <- table(dataset_for_mcnemar[[var1]], dataset_for_mcnemar[[var2]])
  
  # Проверка, что таблица корректна (без пустых клеток)
  if (any(tbl < 1)) {
    return(data.frame(
      variable1 = var1,
      variable2 = var2,
      p.value = NA,
      statistic = NA,
      warning = "Table contains zeros"
    ))
  }
  
  # Выполнение теста Мак-Нимара (пробуем с обработкой исключений)
  test_result <- tryCatch(
    mcnemar.test(tbl),
    error = function(e) {
      return(data.frame(
        variable1 = var1,
        variable2 = var2,
        p.value = NA,
        statistic = NA,
        warning = "Test failed"
      ))
    }
  )

  # Если тест прошел успешно, возвращаем результат
  if (inherits(test_result, "htest")) {
    return(data.frame(
      variable1 = var1,
      variable2 = var2,
      p.value = round(test_result$p.value,5),
      statistic = round(test_result$statistic,5),
      warning = NA
    ))
  } else {
    return(test_result)
  }
}

# Генерация всех возможных пар переменных
all_pairs <- combn(var_for_mcnemar, 2, simplify = FALSE)

# Применение теста Мак-Нимара ко всем парам переменных
results <- map_dfr(all_pairs, ~ run_mcnemar_test_pair(.x[1], .x[2]))

# Вывод результатов
results %>% flextable() %>% 
 theme_box() %>% 
  align(align = "center", part = "all") %>% 
  merge_v("variable1") %>%  save_as_html( values = NULL, path ="data/raw/ результаты_теста_макнеймара.html", lang = "en", title = "&#32;")
```




```{r}

# Задаем желаемый порядок переменных
var_order <- c(
  "memoryloss_v8",
  "moca_bin",
  "mmse_bin",
  "fab_bin",
  "symboldigit_bin",
  "bdi_bin",
  "stsutiational_bin",
  "stpersonal_bin"
)

# Преобразуем данные и устанавливаем конкретный порядок
heatmap_data <- results %>%
# filter(!is.na(p.value)) %>%  # Исключаем невалидные тесты
  select(variable1, variable2, p.value) %>%
  mutate(significant = ifelse(p.value < 0.05, "yes", "no")) %>%  # Добавляем признак значимости
  mutate(
    variable1 = factor(variable1, levels = var_order),  # Упорядочиваем variable1
    variable2 = factor(variable2, levels = var_order)   # Упорядочиваем variable2
  )

# Построение тепловой карты
ggplot(heatmap_data, aes(x = variable1, y = variable2, fill = p.value)) +
  geom_tile(color = "black") + 
  # scale_fill_gradientn(colours = grDevices::Cyan-Magenta(10))+
  scale_fill_gradient(low = "#DEB0D0", high = "white", na.value = "grey90") + # Цветовая шкала
  geom_text(aes(label = gtsummary::style_pvalue(p.value, 3)), color = "black", size = 3) +  # Подписи p-value
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        axis.title = element_blank())+
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +  # Поворот подписи оси X
  labs(title = "Результаты теста Мак-Неймара, p-значения")
```


## Visualisation boxplots
```{r Visualisation_boxplots, fig.height=16, fig.width=30}



 



data_spagetty_before_imp <- vis_all %>% select(id, starts_with(c("bmi",  "athero_i", "sad", "pulse"))) %>% 
  pivot_longer(
    cols = -c(id),
    names_pattern = "(.+)_(.+)",
    names_to = c(".value", "id_point")
  )


# BMI 
boxplot_BMI_before_imp <-
  data_spagetty_before_imp %>%
  filter(!is.na(bmi)) %>%
  ggplot(aes(x = id_point, y = bmi)) +
  geom_line(aes(group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#DEB0D0",
    alpha = 0.8,
    outlier.size = 0.5  ) +
  labs(title = "BMI dynamics by visit before data imputation")

boxplot_BMI_before_imp

ggsave("data/pics/boxplot_BMI_before_imp.png", plot = boxplot_BMI_before_imp,height = 8, width = 12)


# SAD_before_imp
boxplot_SAD_before_imp <-  data_spagetty_before_imp %>%
  filter(!is.na(sad)) %>%
  ggplot(aes(x = id_point, y = sad)) +
  geom_line(aes( group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#DEB0D0",
    alpha = 0.8,
    outlier.size = 0.5
  ) +
  labs(title = "SAD dynamics by visit before data imputation")

boxplot_SAD_before_imp

ggsave(
  "data/pics/boxplot_SAD_before_imp.png",
  plot = boxplot_SAD_before_imp,
  height = 8,
  width = 12
)


# athero_index_before_imp
boxplot_athero_index_before_imp <-  data_spagetty_before_imp %>%
  filter(!is.na(athero_index)) %>%
  ggplot(aes(x = id_point, y = athero_index)) +
  geom_line(aes( group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#DEB0D0",
    alpha = 0.8,
    outlier.size = 0.5
  ) +
  labs(title = "Atherogenic index dynamics by visit before data imputation")

boxplot_athero_index_before_imp

ggsave(
  "data/pics/boxplot_athero_index_before_imp.png",
  plot = boxplot_athero_index_before_imp,
  height = 8,
  width = 12
)









imputed_data_longest_mean <- read_rds( "data/raw/imputed_data_longest_mean.rds") 

# BMI_after_imp
boxplot_BMI_after_imp <-
  imputed_data_longest_mean %>%
  filter(!is.na(bmi)) %>%
  ggplot(aes(x = id_point, y = bmi)) +
  geom_line(aes(group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#86CA78",
    alpha = 0.8,
    outlier.size = 0.5  ) +
  labs(title = "BMI dynamics by visit after data imputation")

boxplot_BMI_after_imp

ggsave("data/pics/boxplot_BMI_after_imp.png", plot = boxplot_BMI_after_imp,height = 8, width = 12)


# SAD_after_imp
boxplot_SAD_after_imp <-  imputed_data_longest_mean %>%
  filter(!is.na(sad)) %>%
  ggplot(aes(x = id_point, y = sad)) +
  geom_line(aes( group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#86CA78",
    alpha = 0.8,
    outlier.size = 0.5
  ) +
  labs(title = "SAD dynamics by visit after data imputation")

boxplot_SAD_after_imp

ggsave(
  "data/pics/boxplot_SAD_after_imp.png",
  plot = boxplot_SAD_after_imp,
  height = 8,
  width = 12
)


# athero_index_after_imp
boxplot_athero_index_after_imp <-  imputed_data_longest_mean %>%
  filter(!is.na(athero_index)) %>%
  ggplot(aes(x = id_point, y = athero_index)) +
  geom_line(aes( group = id),
            col = "grey50",
            alpha = 0.4) +
  theme_custom  +
  geom_boxplot(
    fill = "#86CA78",
    alpha = 0.8,
    outlier.size = 0.5
  ) +
  labs(title = "Atherogenic index dynamics by visit after data imputation")

boxplot_athero_index_after_imp

ggsave(
  "data/pics/boxplot_athero_index_after_imp.png",
  plot = boxplot_athero_index_after_imp,
  height = 8,
  width = 12
)

# Все графики вместе
boxplot_all <- ggpubr::ggarrange(
  plotlist = list(
    boxplot_BMI_before_imp,
    boxplot_SAD_before_imp,
    boxplot_athero_index_before_imp ,
    boxplot_BMI_after_imp,
    boxplot_SAD_after_imp ,
    boxplot_athero_index_after_imp 
  ),
  ncol = 3,
  nrow = 2,
  heights = c(1, 1)
)

boxplot_all

ggsave(
  "data/pics/boxplot_all.png",
  plot = boxplot_all,
  height = 20,
  width = 30
)



```

## Visualisation BARplot
```{r Visualisation BARplot, fig.height=16, fig.width=24}


# барплоты для курильщиков и алкоголиков ДО ИМПУТАЦИИ
data_barplot_before_imp <- vis_all %>% select(id, starts_with(c("smoki", "alco"))) %>% 
  pivot_longer(
    cols = -c(id),
    names_pattern = "(.+)_(.+)",
    names_to = c(".value", "id_point")
  )






# # Подсчет количества значений smoking, разбитых по группам id_point
# data_counts <-
#   data_barplot_before_imp %>% 
#   filter(!is.na(smoking)) %>%
#   group_by(id_point, smoking) %>%
#   summarise(count = n(), .groups = "drop")
# 
# # Визуализация данных
# data_counts %>%
#   ggplot() +
#   geom_col(aes(x = id_point, y = count, fill = smoking), position = "dodge") +
#   labs(
#     x = "id_point",
#     
#     title = "Smoking dynamics by visit before data imputation"
#   ) +
#   theme_custom


##### SMOKING

# Подсчет долей smoking, разбитых по группам id_point
data_ratios_smok <- data_barplot_before_imp %>%
  filter(!is.na(smoking)) %>%
  group_by(id_point, smoking) %>%
  summarise(count = n(), .groups = "drop") %>%
  group_by(id_point) %>%
  mutate(ratio = count / sum(count)) %>%  # Вычисляем долю
mutate(smoking = factor(smoking, levels = c("Non-smoker", "Current smoker"))) # для порядка столбцов

# график с долями
bar_smoking_before_imp <- data_ratios_smok %>%
  ggplot() +
  geom_col(aes(x = id_point, y = ratio, fill = smoking),col="black", position = "dodge") +
  labs(
    x = "id_point",
    y = "ratios",
    title = "Smoking dynamics by visit before data imputation (ratios)"
  ) +
  scale_y_continuous(labels = scales::percent) + 
  scale_fill_manual(values = c("Current smoker" = "#DEB0D0",
                               "Non-smoker" = "#86CA78"))+
    theme_custom

#########################
# ALCOHOL

data_ratios_alc <- data_barplot_before_imp %>%
  filter(!is.na(alcohol)) %>%
  group_by(id_point, alcohol) %>%
  summarise(count = n(), .groups = "drop") %>%
  group_by(id_point) %>%
  mutate(ratio = count / sum(count)) %>%  # Вычисляем долю
mutate(alcohol = factor(alcohol, levels = c("No", "Yes"))) # для порядка столбцов

# график с долями
bar_alcohol_before_imp <- data_ratios_alc %>%
  ggplot() +
  geom_col(aes(x = id_point, y = ratio, fill = alcohol),col="black", position = "dodge") +
  labs(
    x = "id_point",
    y = "ratios",
    title = "Alcohol dynamics by visit before data imputation (ratios)"
  ) +
  scale_y_continuous(labels = scales::percent, limits = c(0,1)) +
  scale_fill_manual(values = c("Yes" = "#DEB0D0",
                               "No" = "#86CA78"))+
    theme_custom



###############
# барплоты для курильщиков и алкоголиков ПОСЛЕ ИМПУТАЦИИ

# imputed_data_longest_mean

##### SMOKING

# Подсчет долей smoking, разбитых по группам id_point
data_ratios_smok_aft <- imputed_data_longest_mean %>%
  filter(!is.na(smoking)) %>%
  group_by(id_point, smoking) %>%
  summarise(count = n(), .groups = "drop") %>%
  group_by(id_point) %>%
  mutate(ratio = count / sum(count)) %>%  # Вычисляем долю
  mutate(smoking = factor(smoking, levels = c("Non-smoker", "Current smoker"))) # для порядка столбцов %>%

# график с долями
bar_smoking_aft_imp <- data_ratios_smok_aft %>%
  ggplot() +
  geom_col(aes(x = id_point, y = ratio, fill = smoking), col="black",position = "dodge") +
  labs(
    x = "id_point",
    y = "ratios",
    title = "Smoking dynamics by visit after data imputation (ratios)"
  ) +
  scale_y_continuous(labels = scales::percent) + 
  scale_fill_manual(values = c("Current smoker" = "#DEB0D0",
                               "Non-smoker" = "#86CA78"))+
    theme_custom


#########################
# ALCOHOL

data_ratios_alc_aft <- imputed_data_longest_mean %>%
  filter(!is.na(alcohol)) %>%
  group_by(id_point, alcohol) %>%
  summarise(count = n(), .groups = "drop") %>%
  group_by(id_point) %>%
  mutate(ratio = count / sum(count)) %>%   # Вычисляем долю
mutate(alcohol = factor(alcohol, levels = c("No", "Yes"))) # для порядка столбцов

# график с долями
bar_alcohol_aft_imp <- data_ratios_alc_aft  %>%
  ggplot() +
  geom_col(aes(x = id_point, y = ratio, fill = alcohol), col="black",position = "dodge") +
  labs(
    x = "id_point",
    y = "ratios",
    title = "Alcohol dynamics by visit after data imputation (ratios)"
  ) +
  scale_y_continuous(labels = scales::percent, limits = c(0,1)) +
  scale_fill_manual(values = c("Yes" = "#DEB0D0",
                               "No" = "#86CA78"))+
    theme_custom


# Все графики вместе
boxplot_all <- ggpubr::ggarrange(
  plotlist = list(
    bar_smoking_before_imp,
    bar_smoking_aft_imp,
    bar_alcohol_before_imp,
    bar_alcohol_aft_imp
  ),
  ncol = 2,
  nrow = 2,
  heights = c(1, 1)
)

boxplot_all

ggsave(
  "data/pics/barplot_all.png",
  plot = boxplot_all,
  height = 16,
  width = 24
)


```

## Оценка согласованности шкал тестов (Тесты для оценки состояния когнитивной сферы)
!!! МЕТОД НЕ ПОДХОДИТ из-за разности шкал

```{r bland.altman.plot.cognitive, fig.height= 10, fig.width=20}
# Проверяем согласованость четырех шкал.
# 1. Краткая шкала оценки психического статуса (КШОПС) / Mini-mental State Examination (MMSE) – скрининговая переменная ‘mmse’
# 2.	Монреальская шкала когнитивной оценки (МоСА) – скрининговая, более чувствительная к когнитивным нарушениям сосудистой этиологии  переменная ‘moca’
# 3.	FAB (Frontal Assessment Batter - "Батарея лобной дисфункции") – выявляет нарушения лобных долей и их связи с подкорковыми структурами переменная ‘fab’
# 4.	Тест символьно-цифрового кодирования – оценка скорости мыслительных процессов и концентрации внимания (работа теменной и затылочной долей) переменная ‘symboldigit’

# Проверка будет осуществлятся следующими парами
# moca ~ mmse, moca ~ fab, moca ~ symboldigit,  mmse ~ fab, mmse ~ symboldigit,  fab ~ symboldigit


vis_all %>% select(starts_with(c("id", "moca_v8",  "mmse_v8", "fab_v8", "symboldigit_v8")))


MOCA_vs_MMSE <-  bland.altman.plot(
  vis_all$moca_v8,
  vis_all$mmse_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "MOCA vs MMSE",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))

 

MOCA_vs_FAB <-  bland.altman.plot(
  vis_all$moca_v8,
  vis_all$fab_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "MOCA vs FAB",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))


  
MOCA_vs_symboldigit <-  bland.altman.plot(
  vis_all$moca_v8,
  vis_all$symboldigit_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "MOCA vs symboldigit",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))


MMSE_vs_FAB <-  bland.altman.plot(
  vis_all$mmse_v8,
  vis_all$fab_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "MMSE vs FAB",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))

 

MMSE_vs_symboldigit <-  bland.altman.plot(
  vis_all$mmse_v8,
  vis_all$symboldigit_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "MMSE vs symboldigit",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))



FAB_vs_symboldigit <-  bland.altman.plot(
  vis_all$fab_v8,
  vis_all$symboldigit_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "FAB vs symboldigit",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))





p <- ggpubr::ggarrange(plotlist = list(MOCA_vs_MMSE, MOCA_vs_FAB, MOCA_vs_symboldigit, MMSE_vs_FAB , MMSE_vs_symboldigit, FAB_vs_symboldigit),
                  ncol = 3,
          nrow = 2,  
          heights = c(1,1),
          label.y = 0)

ggsave("bland_altman_plot_cognitive.png", plot = , path = "data/pics/", width = 20,
  height = 10)
```


## Оценка согласованности шкал тестов (Тесты для оценки эмоционально-аффективной сферы)

```{r bland.altman.plot.emotion, fig.height= 6, fig.width=20}
# Проверяем согласованость двух шкал.
# 1.	Шкала депрессии Бека-II  переменная ‘bdi’
# 2.	Шкала тревоги Cпилбергера в адаптации Ханина переменные ‘stsutiational’ и ‘stpersonal’
# 
#  Проверка будет осуществлятся следующими парами
#  bdi ~ stsutiational, bdi ~ stpersonal, stsutiational ~ stpersonal

vis_all %>% select(starts_with(c("id", "bdi_v8", "stsutiational_v8" , "stpersonal_v8")))

BDI_vs_stsutiational <-  bland.altman.plot(
  vis_all$bdi_v8,
  vis_all$stsutiational_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "BDI vs stsutiational",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))

BDI_vs_stsutiational



BDI_vs_stpersonal <-  bland.altman.plot(
  vis_all$bdi_v8,
  vis_all$stpersonal_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "BDI vs stpersonal",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))

BDI_vs_stpersonal



stsutiational_vs_stpersonal <-  bland.altman.plot(
  vis_all$bdi_v8,
  vis_all$stpersonal_v8,
  graph.sys = "ggplot2",
  conf.int = .95,
  geom_count = TRUE,
  silent = TRUE
) +
  labs (title = "stsutiational vs stpersonal",
        # subtitle = ,
        y = "Differences",
        x = "Means of measurements") +
  theme_bw() +
  theme (plot.title = element_text( 
                                    size = 30, 
                                    hjust = 0.5))


stsutiational_vs_stpersonal

p02 <- ggpubr::ggarrange(plotlist = list(BDI_vs_stpersonal, BDI_vs_stsutiational, stsutiational_vs_stpersonal),
                  ncol = 3,
          nrow = 1,  
          heights = c(1,1),
          label.y = 0)
p02

ggsave("bland_altman_plot_emotion.png", plot = p02, path = "data/pics/", width = 20,
  height = 6)

```
