==> 04_Depression_mri_cognition.Rmd <==
---
title: "Assoosiations between depression, anxiety and cognitive impairment"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
library(MASS)
library(psych)
library(ggplot2)
library(dplyr)
library(tidyr)
```

```{r}
#Импорт данных и создание переменной - исхода
data <- readRDS("data/raw/vis_all.rds")
#glimpse(data)



```

```{r}

# Оценка размера каждой группы
data %>% filter(!is.na(data$cg_imp_all)) %>% 
  mutate(cg_imp_all = as.factor(cg_imp_all)) %>%  
  group_by(cg_imp_all) %>%
  summarise(n = n()) %>%
  mutate(percent = n/sum(n) * 100)
```

```{r ANOVA, echo=F}

# Сравнение тревожности идепрессии в группах
model_anxiety_personal <- aov(stpersonal_v8 ~ factor(cg_imp_all), data) 
summary(model_anxiety_personal)
TukeyHSD(model_anxiety_personal)

model_anxiety_sutiational <- aov(stsutiational_v8 ~ factor(cg_imp_all), data) 
summary(model_anxiety_sutiational)
TukeyHSD(model_anxiety_sutiational)

model_bdi <- aov(bdi_v8 ~ factor(cg_imp_all), data) 
summary(model_bdi)
TukeyHSD(model_bdi)

```

```{r SPEARMAN, echo=F}
# Оценка корреляции Спирмана
cor.test(data$bdi_v8, as.numeric(data$cg_imp_all), method="spearman") 
cor.test(data$stsutiational_v8, as.numeric(data$cg_imp_all), method="spearman")
cor.test(data$stpersonal_v8, as.numeric(data$cg_imp_all), method="spearman") 

```

```{r SPEARMAN_VIS, echo=F}
# Визуализация корреляции 
plot_data_cor <- data %>%
  dplyr::select(c("bdi_v8", "stsutiational_v8", "stpersonal_v8", "cg_imp_all")) %>%
  pivot_longer(cols = c("bdi_v8", "stsutiational_v8", "stpersonal_v8"), names_to = "variable", values_to = "value")

ggplot(plot_data_cor, aes(x = value, y = as.numeric(cg_imp_all))) +  
  geom_point(alpha = 0.6) +  # Отображаем точки
  geom_smooth(method = "lm", se = FALSE, color = "blue") +  
  facet_wrap(~variable, scales = "free_x", 
             labeller = labeller(variable = c(
               bdi = "Депрессия",
               stsutiational = "Ситуационная тревожность",
               stpersonal = "Личностная тревожность"
             ))) +  # Добавляем подписи
  labs(
    title = "Корреляции между переменными и категориями когнитивных нарушений",
    x = "Значение переменной",
    y = "Категория когнитивных нарушений (числовое значение)",
    color = "Переменная"
  ) +
  theme_minimal()

```

```{r ORD_LOG_REG, echo=FALSE}
# Порядковая логистическая регрессия  
model <- polr(factor(cg_imp_all) ~ bdi_v8 + stsutiational_v8 + stpersonal_v8, data, Hess=TRUE)
summary(model)
p_values <- 2 * (1 - pnorm(abs(summary(model)$coefficients[, "t value"])))
p_values
confint(model)

```

```{r ORD_LOG_REG_BDI, echo=FALSE}
# Порядковая логистическая регрессия  
model_restricted <- polr( factor(cg_imp_all) ~ bdi_v8, data = data, Hess = TRUE)

summary(model_restricted)

p_values <- 2 * (1 - pnorm(abs(summary(model_restricted)$coefficients[, "t value"])))
p_values

confint(model_restricted)

```

```{r DEPRESSION_VIS, echo=F}
# Визуализация порядковой логистической регрессии
new_data <- data.frame(bdi_v8 = seq(min(data$bdi_v8, na.rm = TRUE), max(data$bdi_v8, na.rm = TRUE), length.out = 100))

predicted_probs <- predict(model_restricted, new_data, type = "probs")

plot_data <- as.data.frame(predicted_probs) %>%
  mutate(bdi = new_data$bdi_v8) %>%
  pivot_longer(cols = -bdi, names_to = "cg_imp_all", values_to = "probability")

plot_data$cg_imp_all <- factor(plot_data$cg_imp_all, levels = colnames(predicted_probs))

labels <- c("0" = "Нет когнитивных нарушений",
            "1" = "Субъективные когнитивные нарушения",
            "2" = "Умеренные когнитивные нарушения",
            "3" = "Не проходил исследование")

ggplot(plot_data, aes(x = bdi, y = probability, color = cg_imp_all)) +
  geom_line(size = 1) +
  labs(
    title = "Предсказанные вероятности категорий когнитивных нарушений \n в зависимости от баллов по шкале депресии Бека",
    x = "Баллы BDI",
    y = "Вероятность",
    color = "Категория когнитивных нарушений"
  ) +
  scale_color_manual(values = c("0" = "blue", "1" = "orange", "2" = "green", "3" = "red"), labels = labels) +
  theme_minimal()

```

```{r MRI_FISHER, echo=F}
# Проверка ассоциаций между биомаркерами МРТ и когнитивными нарушениями
mri_vars <- c("mriseverallesions_v8", "mrimultiplelesions_v8", "mriperivascularspaces_v8",
              "mricsfspaces_v8", "mrilacunes_v8", "mristroke_v8", "mrihippocampus_v8")

p_values <- data.frame(variable = character(), p_value = numeric())

significant_vars <- c()

for (var in mri_vars) {
  print(paste("Тест Фишера для", var))

  table <- table(data[[var]], data$cg_imp_all)

  if (all(dim(table) == c(2, 2))) {
    test_result <- fisher.test(table)
  } else {
        test_result <- chisq.test(table)
  }

  p_val <- test_result$p.value
  p_values <- rbind(p_values, data.frame(variable = var, p_value = p_val))

  if (p_val < 0.05) {
      significant_vars <- c(significant_vars, var)
    }
}


print(p_values)



# mri_vars <- c("mriseverallesions_v8", "mrimultiplelesions_v8", "mriperivascularspaces_v8", 
#               "mricsfspaces_v8", "mrilacunes_v8", "mristroke_v8", "mrihippocampus_v8")
# 
# p_values <- data.frame(variable = character(), p_value = numeric(), stringsAsFactors = FALSE)
# 
# significant_vars <- c()
# 
# for (var in mri_vars) {
#   print(paste("Тест для переменной:", var))
#   
#   # Проверка уникальных значений и удаления NA
#   clean_data <- data[!is.na(data[[var]]) & !is.na(data$cg_imp_all), ]
#   
#   if (length(unique(clean_data[[var]])) > 1) {
#     table <- table(clean_data[[var]], clean_data$cg_imp_all)
#     
#     # Проверяем размер таблицы
#     if (all(dim(table) == c(2, 2))) {
#       test_result <- fisher.test(table)
#     } else if (nrow(table) >= 2 & ncol(table) >= 2) {
#       test_result <- chisq.test(table)
#     } else {
#       print(paste("Переменная", var, "имеет неподходящий размер таблицы:", dim(table)))
#       next
#     }
#     
#     # Сохраняем p-value
#     p_val <- test_result$p.value
#     p_values <- rbind(p_values, data.frame(variable = var, p_value = p_val))
#     
#     # Проверка значимости
#     if (p_val < 0.05) {
#       significant_vars <- c(significant_vars, var)
#     }
#   } else {
#     print(paste("Переменная", var, "имеет слишком мало уровней, пропускается"))
#   }
# }
# 
# # Вывод результатов
# print("Все p-значения:")
# print(p_values)
# 
# if (length(significant_vars) > 0) {
#   print("Значимые переменные:")
#   print(significant_vars)
# } else {
#   print("Нет значимых переменных")
# }
```

```{r ORD_LOG_REG_MRI, echo=F}
# Порядковая логистическая регрессия (миниинсульт)
model_mri <- polr(factor(cg_imp_all) ~ mristroke_v8, data = data, Hess = TRUE)
 
summary(model_mri)

p_values <- 2 * (1 - pnorm(abs(summary(model_mri)$coefficients[, "t value"])))
p_values

confint(model_mri)

```
