==> 07_LMM_analysis.Rmd <==
---
title: "LMM_Analysis"
date: "2025-01-30"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(mice)
library(lme4)
library(lmerTest)
library(dplyr)
library(sjPlot)
library(performance)
library(broom.mixed)
library(purrr)
library(tidyr)
library(flextable)

```

```{r}
mids <- readRDS("data/raw/imputed_data_mids.rds")

#1. `sbp` → `moca_v8`
mod_full_sbp_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, moca_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ moca_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_sbp_moca, mod_null_sbp_moca)


#2. `sbp` → `mmse_v8`
mod_full_sbp_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, mmse_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ mmse_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_mmse, mod_null_sbp_mmse)


#3. `sbp` → `fab_v8`
mod_full_sbp_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, fab_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ fab_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_fab, mod_null_sbp_fab)


#4. `sbp` → `symboldigit_v8`
mod_full_sbp_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, symboldigit_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ symboldigit_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_symboldigit, mod_null_sbp_symboldigit)


#5. `sbp` → `stsutiational_v8`
mod_full_sbp_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, stsutiational_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ stsutiational_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_stsutiational, mod_null_sbp_stsutiational)


#6. `sbp` → `stpersonal_v8`
mod_full_sbp_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, stpersonal_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ stpersonal_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_stpersonal, mod_null_sbp_stpersonal)


#7`sbp` → `bdi_v8`
mod_full_sbp_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id, bdi_v8), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      lmer(sbp ~ bdi_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_sbp_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("sbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "sbp_", 
                   names_to = "visit", 
                   values_to = "sbp") %>% 
      filter(!is.na(sbp)) %>% 
      lmer(sbp ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_sbp_bdi, mod_null_sbp_bdi)

get_d1_results <- function(full_model, null_model, predictor, outcome) {
  d1_result <- D1(full_model, null_model)
  
  if (!is.null(d1_result$result) && all(c("F.value", "df1", "df2", "P(>F)") %in% colnames(d1_result$result))) {
    result <- data.frame(
      predictor = predictor,
      outcome = outcome,
      F_value = d1_result$result[1, "F.value"],
      df1 = d1_result$result[1, "df1"],
      df2 = d1_result$result[1, "df2"],
      p_value = d1_result$result[1, "P(>F)"],
      stringsAsFactors = FALSE
    )
    return(result)
  } else {
    message("Ошибка: некорректная структура d1_result для предиктора ", predictor, " и исхода ", outcome)
    return(NULL)
  }
}

models_list <- list(
  list(mod_full_sbp_moca, mod_null_sbp_moca, "sbp", "moca_v8"),
  list(mod_full_sbp_mmse, mod_null_sbp_mmse, "sbp", "mmse_v8"),
  list(mod_full_sbp_fab, mod_null_sbp_fab, "sbp", "fab_v8"),
  list(mod_full_sbp_symboldigit, mod_null_sbp_symboldigit, "sbp", "symboldigit_v8"),
  list(mod_full_sbp_stsutiational, mod_null_sbp_stsutiational, "sbp", "stsutiational_v8"),
  list(mod_full_sbp_stpersonal, mod_null_sbp_stpersonal, "sbp", "stpersonal_v8"),
  list(mod_full_sbp_bdi, mod_null_sbp_bdi, "sbp", "bdi_v8")
)

results_sbp <- models_list %>%
  map_dfr(~ {
    tryCatch({
      get_d1_results(.x[[1]], .x[[2]], .x[[3]], .x[[4]])
    }, error = function(e) {
      message("Ошибка для предиктора ", .x[[3]], " и исхода ", .x[[4]], ": ", e$message)
      return(NULL)
    })
  })

ft_sbp <- results_sbp %>%
  flextable() %>%
  set_header_labels(
    predictor = "Predictor",
    outcome = "Outcome",
    F_value = "F_value",
    df1 = "Degrees of freedom 1",
    df2 = "Degrees of freedom 1",
    p_value = "p-value"
  ) %>%
  colformat_double(j = c("F_value", "df1", "df2"), digits = 2) %>%
  colformat_double(j = "p_value", digits = 3) %>%
  theme_zebra() %>%
  autofit()

ft_sbp
```

```{r}
# 1. `dbp` → `moca_v8`
mod_full_dbp_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, moca_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ moca_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_moca, mod_null_dbp_moca)

# 2. `dbp` → `mmse_v8`
mod_full_dbp_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, mmse_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ mmse_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_mmse, mod_null_dbp_mmse)

# 3. `dbp` → `fab_v8`
mod_full_dbp_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, fab_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ fab_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_fab, mod_null_dbp_fab)

# 4. `dbp` → `symboldigit_v8`
mod_full_dbp_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, symboldigit_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ symboldigit_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_symboldigit, mod_null_dbp_symboldigit)

# 5. `dbp` → `stsutiational_v8`
mod_full_dbp_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, stsutiational_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ stsutiational_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_stsutiational, mod_null_dbp_stsutiational)

# 6. `dbp` → `stpersonal_v8`
mod_full_dbp_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, stpersonal_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ stpersonal_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_stpersonal, mod_null_dbp_stpersonal)

# 7. `dbp` → `bdi_v8`
mod_full_dbp_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id, bdi_v8), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      lmer(dbp ~ bdi_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_dbp_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("dbp_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "dbp_", 
                   names_to = "visit", 
                   values_to = "dbp") %>% 
      filter(!is.na(dbp)) %>% 
      lmer(dbp ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_dbp_bdi, mod_null_dbp_bdi)

models_list_dbp <- list(
  list(mod_full_dbp_moca, mod_null_dbp_moca, "dbp", "moca_v8"),
  list(mod_full_dbp_mmse, mod_null_dbp_mmse, "dbp", "mmse_v8"),
  list(mod_full_dbp_fab, mod_null_dbp_fab, "dbp", "fab_v8"),
  list(mod_full_dbp_symboldigit, mod_null_dbp_symboldigit, "dbp", "symboldigit_v8"),
  list(mod_full_dbp_stsutiational, mod_null_dbp_stsutiational, "dbp", "stsutiational_v8"),
  list(mod_full_dbp_stpersonal, mod_null_dbp_stpersonal, "dbp", "stpersonal_v8"),
  list(mod_full_dbp_bdi, mod_null_dbp_bdi, "dbp", "bdi_v8")
)

results_dbp <- models_list_dbp %>%
  map_dfr(~ {
    tryCatch({
      get_d1_results(.x[[1]], .x[[2]], .x[[3]], .x[[4]])
    }, error = function(e) {
      message(" Ошибка для предиктора ", .x[[3]], " и исхода ", .x[[4]], ": ", e$message)
      return(NULL)
    })
  })

ft_dbp <- results_dbp %>%
  flextable() %>%
  set_header_labels(
    predictor = "Predictor",
    outcome = "Outcome",
    F_value = "F_value",
    df1 = "Degrees of freedom 1",
    df2 = "Degrees of freedom 2",
    p_value = "p-value"
  ) %>%
  colformat_double(j = c("F_value", "df1", "df2"), digits = 2) %>%
  colformat_double(j = "p_value", digits = 3) %>%
  theme_zebra() %>%
  autofit()

ft_dbp
```


```{r}
# 1. `pulse` → `moca_v8`
mod_full_pulse_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, moca_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ moca_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_moca, mod_null_pulse_moca)

# 2. `pulse` → `mmse_v8`
mod_full_pulse_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, mmse_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ mmse_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_mmse, mod_null_pulse_mmse)

# 3. `pulse` → `fab_v8`
mod_full_pulse_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, fab_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ fab_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_fab, mod_null_pulse_fab)

# 4. `pulse` → `symboldigit_v8`
mod_full_pulse_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, symboldigit_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ symboldigit_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_symboldigit, mod_null_pulse_symboldigit)

# 5. `pulse` → `stsutiational_v8`
mod_full_pulse_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, stsutiational_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ stsutiational_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_stsutiational, mod_null_pulse_stsutiational)

# 6. `pulse` → `stpersonal_v8`
mod_full_pulse_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, stpersonal_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ stpersonal_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_stpersonal, mod_null_pulse_stpersonal)

# 7. `pulse` → `bdi_v8`
mod_full_pulse_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id, bdi_v8), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      lmer(pulse ~ bdi_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_pulse_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("pulse_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "pulse_", 
                   names_to = "visit", 
                   values_to = "pulse") %>% 
      filter(!is.na(pulse)) %>% 
      lmer(pulse ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_pulse_bdi, mod_null_pulse_bdi)

models_list_pulse <- list(
  list(mod_full_pulse_moca, mod_null_pulse_moca, "pulse", "moca_v8"),
  list(mod_full_pulse_mmse, mod_null_pulse_mmse, "pulse", "mmse_v8"),
  list(mod_full_pulse_fab, mod_null_pulse_fab, "pulse", "fab_v8"),
  list(mod_full_pulse_symboldigit, mod_null_pulse_symboldigit, "pulse", "symboldigit_v8"),
  list(mod_full_pulse_stsutiational, mod_null_pulse_stsutiational, "pulse", "stsutiational_v8"),
  list(mod_full_pulse_stpersonal, mod_null_pulse_stpersonal, "pulse", "stpersonal_v8"),
  list(mod_full_pulse_bdi, mod_null_pulse_bdi, "pulse", "bdi_v8")
)

results_pulse <- models_list_pulse %>%
  map_dfr(~ {
    tryCatch({
      get_d1_results(.x[[1]], .x[[2]], .x[[3]], .x[[4]])
    }, error = function(e) {
      message("Ошибка для предиктора ", .x[[3]], " и исхода ", .x[[4]], ": ", e$message)
      return(NULL)
    })
  })

ft_pulse <- results_pulse %>%
  flextable() %>%
  set_header_labels(
    predictor = "Predictor",
    outcome = "Outcome",
    F_value = "F_value",
    df1 = "Degrees of freedom 1",
    df2 = "Degrees of freedom 2",
    p_value = "p-value"
  ) %>%
  colformat_double(j = c("F_value", "df1", "df2"), digits = 2) %>%
  colformat_double(j = "p_value", digits = 3) %>%
  theme_zebra() %>%
  autofit()

ft_pulse

```

```{r}

# 1. `athero_index` → `moca_v8`
mod_full_athero_index_moca<- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, moca_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>% 
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>% 
      lmer(athero_index ~ moca_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_moca <-  mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>% 
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>% 
      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })

D1(mod_full_athero_index_moca, mod_null_athero_index_moca)

# 2. `athero_index` → `mmse_v8`
mod_full_athero_index_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, mmse_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ mmse_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>% 
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>% 
      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_mmse, mod_null_athero_index_mmse)

# 3. `athero_index` → `fab_v8`
mod_full_athero_index_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, fab_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ fab_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_fab, mod_null_athero_index_fab)

# 4. `athero_index` → `symboldigit_v8`
mod_full_athero_index_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, symboldigit_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ symboldigit_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>% 
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_symboldigit, mod_null_athero_index_symboldigit)

# 5. `athero_index` → `stsutiational_v8`
mod_full_athero_index_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, stsutiational_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ stsutiational_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_stsutiational, mod_null_athero_index_stsutiational)

# 6. `athero_index` → `stpersonal_v8`
mod_full_athero_index_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, stpersonal_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ stpersonal_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_stpersonal, mod_null_athero_index_stpersonal)

# 7. `athero_index` → `bdi_v8`
mod_full_athero_index_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id, bdi_v8), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%
      lmer(athero_index ~ bdi_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_athero_index_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("athero_index_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "athero_index_", 
                   names_to = "visit", 
                   values_to = "athero_index") %>%
      filter(!is.infinite(athero_index) & !is.na(athero_index) & !is.nan(athero_index)) %>%      lmer(athero_index ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_athero_index_bdi, mod_null_athero_index_bdi)

models_list_athero_index <- list(
  list(mod_full_athero_index_moca, mod_null_athero_index_moca, "athero_index", "moca_v8"),
  list(mod_full_athero_index_mmse, mod_null_athero_index_mmse, "athero_index", "mmse_v8"),
  list(mod_full_athero_index_fab, mod_null_athero_index_fab, "athero_index", "fab_v8"),
  list(mod_full_athero_index_symboldigit, mod_null_athero_index_symboldigit, "athero_index", "symboldigit_v8"),
  list(mod_full_athero_index_stsutiational, mod_null_athero_index_stsutiational, "athero_index", "stsutiational_v8"),
  list(mod_full_athero_index_stpersonal, mod_null_athero_index_stpersonal, "athero_index", "stpersonal_v8"),
  list(mod_full_athero_index_bdi, mod_null_athero_index_bdi, "athero_index", "bdi_v8")
)

results_athero_index <- models_list_athero_index %>%
  map_dfr(~ {
    tryCatch({
      get_d1_results(.x[[1]], .x[[2]], .x[[3]], .x[[4]])
    }, error = function(e) {
      message("Ошибка для предиктора ", .x[[3]], " и исхода ", .x[[4]], ": ", e$message)
      return(NULL)
    })
  })

ft_athero_index <- results_athero_index %>%
  flextable() %>%
  set_header_labels(
    predictor = "Predictor",
    outcome = "Outcome",
    F_value = "F_value",
    df1 = "Degrees of freedom 1",
    df2 = "Degrees of freedom 2",
    p_value = "p-value"
  ) %>%
  colformat_double(j = c("F_value", "df1", "df2"), digits = 2) %>%
  colformat_double(j = "p_value", digits = 3) %>%
  theme_zebra() %>%
  autofit()

ft_athero_index

```

```{r}
# 1. `bmi` → `moca_v8`
mod_full_bmi_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, moca_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ moca_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_moca <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, moca_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_moca, mod_null_bmi_moca)

# 2. `bmi` → `mmse_v8`
mod_full_bmi_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, mmse_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ mmse_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_mmse <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, mmse_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_mmse, mod_null_bmi_mmse)

# 3. `bmi` → `fab_v8`
mod_full_bmi_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, fab_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ fab_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_fab <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, fab_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_fab, mod_null_bmi_fab)

# 4. `bmi` → `symboldigit_v8`
mod_full_bmi_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, symboldigit_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ symboldigit_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_symboldigit <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, symboldigit_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_symboldigit, mod_null_bmi_symboldigit)

# 5. `bmi` → `stsutiational_v8`
mod_full_bmi_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, stsutiational_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ stsutiational_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_stsutiational <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stsutiational_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_stsutiational, mod_null_bmi_stsutiational)

# 6. `bmi` → `stpersonal_v8`
mod_full_bmi_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, stpersonal_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ stpersonal_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_stpersonal <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, stpersonal_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_stpersonal, mod_null_bmi_stpersonal)

# 7. `bmi` → `bdi_v8`
mod_full_bmi_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id, bdi_v8), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      lmer(bmi ~ bdi_v8 * visit + (1 | id), ., REML = FALSE)
  })

mod_null_bmi_bdi <- mice::complete(mids, "all") %>% 
  map(function(df) {
    df %>% 
      select(id, bdi_v8, starts_with("bmi_")) %>% 
      pivot_longer(cols = -c(id), 
                   names_prefix = "bmi_", 
                   names_to = "visit", 
                   values_to = "bmi") %>% 
      filter(!is.na(bmi)) %>% 
      lmer(bmi ~ visit + (1 | id), ., REML = FALSE)
  })
D1(mod_full_bmi_bdi, mod_null_bmi_bdi)

models_list_bmi <- list(
  list(mod_full_bmi_moca, mod_null_bmi_moca, "bmi", "moca_v8"),
  list(mod_full_bmi_mmse, mod_null_bmi_mmse, "bmi", "mmse_v8"),
  list(mod_full_bmi_fab, mod_null_bmi_fab, "bmi", "fab_v8"),
  list(mod_full_bmi_symboldigit, mod_null_bmi_symboldigit, "bmi", "symboldigit_v8"),
  list(mod_full_bmi_stsutiational, mod_null_bmi_stsutiational, "bmi", "stsutiational_v8"),
  list(mod_full_bmi_stpersonal, mod_null_bmi_stpersonal, "bmi", "stpersonal_v8"),
  list(mod_full_bmi_bdi, mod_null_bmi_bdi, "bmi", "bdi_v8")
)

results_bmi <- models_list_bmi %>%
  map_dfr(~ {
    tryCatch({
      get_d1_results(.x[[1]], .x[[2]], .x[[3]], .x[[4]])
    }, error = function(e) {
      message("Ошибка для предиктора ", .x[[3]], " и исхода ", .x[[4]], ": ", e$message)
      return(NULL)
    })
  })

ft_bmi <- results_bmi %>%
  flextable() %>%
  set_header_labels(
    predictor = "Predictor",
    outcome = "Outcome",
    F_value = "F_value",
    df1 = "Degrees of freedom 1",
    df2 = "Degrees of freedom 2",
    p_value = "p-value"
  ) %>%
  colformat_double(j = c("F_value", "df1", "df2"), digits = 2) %>%
  colformat_double(j = "p_value", digits = 3) %>%
  theme_zebra() %>%
  autofit()

ft_bmi

```


```{r}
tables <- list(
  ft_athero_index = ft_athero_index,
  ft_bmi = ft_bmi,
  ft_dbp = ft_dbp,
  ft_pulse = ft_pulse,
  ft_sbp = ft_sbp
)

walk2(names(tables), tables, ~ save_as_image(.y, path = paste0(.x, ".png"), res = 300))

```
